Add PulseAudio output support
This is an initial implementation of a PulseAudio audio engine. It
follows what appear to be considered best practices & configuration for
low audio latency and seems to work well from some personal testing.

Compared to the recommended practices, the main distinction is that the
engine will not attempt to automatically scale the latency, instead
notifying the user so they can adjust the buffer size themselves. This
seems to be more in line with the general style Zrythm uses.

The code was closely inspired by the SDL, JACK, and PortAudio engines. I
also tried to stay as close to the existing code style as possible.