wait for all dsp threads to become idle on startup
when activating the engine, block and wait for all DSP threads to go in use before proceeding. this fixes a bug where if for some reason thread initialization is slow and the threads are not initialized in time before the graph is terminated there is a mismatch in the expected number of threads to join.