fix error when loading project with different samplerate
The audio pool was getting initialized before setting up the audio
engine (where the sample rate is fetched from the backend) so the
audio clips were initialized with whatever the project's sample rate
was.

Made it so that the audio pool is re-initialized after the engine is
set up.

Fixes https://todo.sr.ht/~alextee/zrythm-bug/526.