fix error when attempting to open plugins without carla
If Zrythm was run before built with carla and the plugin setting was
saved as using carla to open, and then Zrythm is run without carla
support and the same plugin is attempted to be open, an error occured.
This fixes it by checking if carla support is present when cloning a
plugin setting and disabling open_with_carla if not.

Fixes https://todo.sr.ht/~alextee/zrythm-bug/531.