work around pipewire bug and fix deadlock in engine
The deadlock was occurring because engine_wait_for_pause() was
waiting for the monitor fade out to end but processing was not
going on so it never occurred. Added a timeout to abort it.

To work around the pipewire bug where it doesn't inform the host
about buffer size changes, the buffer size is checked on every
cycle for changes.

Fixes https://todo.sr.ht/~alextee/zrythm-bug/1054.