don't restart the engine when queueing samples for playback
Create an extra graph for the sample processor instead of pausing the
engine and re-creating the main graph. This removes the lag and fade
in/out when restarting the audio engine and fixes the attack part of
samples being faded-in and inaudible.

A side effect is that the engine will skip its cycle while the sample
processor graph is being recreated so there may be clicks if something
is playing in the timeline.

Fixes #3323.