move zoom controls and event toggle to end of editor toolbar
This is achieved by wrapping the toolbar inside a `GtkCenterBox`. This
also makes it easy to add buttons to the `center` in the future.

The `start` box contains a `GtkScrolledWindow`, to enable smaller
window sizes.

The ghost note toggle is kept at `start`, its named separator removed
as no entry follows it.

Signed-off-by: Miró Allard <miro.allard@pm.me>