refactor project graph generation
* Add ProcessorBase class for all processors with ports & parameters
* Use ProcessorBase to simplify graph creation
* Make standard connections during graph generation instead of keeping
  track of connections with PortConnectionsManager
* Add ChannelSubgraphBuilder to help with generation of sub-graphs for
  each channel
* Clean up now-unused/unneeded API

Still need to fix a few more components to follow the new way of
generating the graph.

Idea for the future: introduce the notion of "taps" into the graph where
we are interested for signals and add ability to "prune" the graph so
that only the access points we are interested in remain.